fixed ‘unable to run test server’
